+++ differt2d/defaults.py
@@ -1,13 +1,7 @@
 """Default values."""
 
-from typing import Literal
-
 DEFAULT_ALPHA: float = 100.0
 """Default value for ``alpha`` parameter in
-:func:`activation<differt2d.logic.activation>`."""
-
-DEFAULT_FUNCTION: Literal["sigmoid", "hard_sigmoid"] = "hard_sigmoid"
-"""Default value for ``function`` parameter in
 :func:`activation<differt2d.logic.activation>`."""
 
 DEFAULT_PATCH: float = 0.0